# _archives/M1 - Web full stack/2025-11-07 - Guide Web API ASP.NET Core C#/MyWebAPI/Models/Health.cs
namespace MyWebAPI.Models
{
    public class Health
    {
        public string Status { get; set; } = "Healthy";
        public DateTime Timestamp { get; set; } = DateTime.UtcNow;
        public string? Version { get; set; }
        public Dictionary<string, object>? Details { get; set; }
    }
}