# _archives/M1 - Web full stack/2025-11-07 - Guide Web API ASP.NET Core C#/MyWebAPI/Models/WeatherForecast.cs
namespace MyWebAPI.Models;

/// <summary>
/// Represents a weather forecast for a single day.
/// </summary>
public record WeatherForecast(
    DateTime Date,
    int TemperatureC,
    string? Summary)
{
    /// <summary>
    /// Temperature in Fahrenheit (computed).
    /// </summary>
    public int TemperatureF => 32 + (int)(TemperatureC * 9.0 / 5.0);
}