# _archives/M1 - Web full stack/2025-11-07 - Guide Web API ASP.NET Core C#/MyWebAPI/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MyWebAPI - User Management System",
    "description": "\n                        <h2>🚀 ASP.NET Core 8.0 Web API</h2>\n                        <p>Complete REST API for managing users with CRUD operations.</p>\n                        \n                        <h3>✨ Features</h3>\n                        <ul>\n                            <li>✅ Full CRUD operations (Create, Read, Update, Delete)</li>\n                            <li>📊 Business analytics (Average Age calculation)</li>\n                            <li>🔒 Data validation with error handling</li>\n                            <li>📚 Complete XML documentation</li>\n                        </ul>\n                        \n                        <h3>🌐 Base URL</h3>\n                        <code>http://localhost:5137</code>\n                    ",
    "contact": {
      "name": "Your Name",
      "url": "https://github.com/yourusername",
      "email": "your.email@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "paths": {
    "/Health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Basic health check endpoint.\r\nReturns API status with uptime and version information.",
        "description": "GET /Health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/Health/detailed": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Detailed health check with full system information.",
        "description": "GET /Health/detailed",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/Health/ping": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Simple ping endpoint for connectivity testing.",
        "description": "GET /Health/ping",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieves all users from the system.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> GET\r\n\n\r\n  <strong>Route:</strong> /Users\r\n\n\r\n  <strong>Authentication:</strong> None required (public endpoint)\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            GET http://localhost:5137/Users\r\n            Accept: application/json\r\n            ```\r\n\n\r\n  <strong>Example Response (200 OK):</strong>\r\n\r\n```\r\n            [\r\n              {\r\n                \"id\": 1,\r\n                \"name\": \"John Doe\",\r\n                \"email\": \"john@example.com\",\r\n                \"age\": 30,\r\n                \"createdAt\": \"2025-11-07T10:46:00Z\",\r\n                \"updatedAt\": null\r\n              },\r\n              {\r\n                \"id\": 2,\r\n                \"name\": \"Jane Smith\",\r\n                \"email\": \"jane@example.com\",\r\n                \"age\": 25,\r\n                \"createdAt\": \"2025-11-07T10:46:00Z\",\r\n                \"updatedAt\": \"2025-11-07T11:30:00Z\"\r\n              }\r\n            ]\r\n            ```",
        "responses": {
          "200": {
            "description": "Successfully returns the list of users (may be empty if no users exist)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Creates a new user in the system.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> POST\r\n\n\r\n  <strong>Route:</strong> /Users\r\n\n\r\n  <strong>Content-Type:</strong> application/json\r\n\n\r\n  <strong>Validation Rules:</strong>\r\n\r\n<list type=\"bullet\">\r\n  <item>Name: Required, 2-100 characters</item>\r\n  <item>Email: Required, valid email format, must be unique</item>\r\n  <item>Age: Required, between 1 and 150</item>\r\n</list>\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            POST http://localhost:5137/Users\r\n            Content-Type: application/json\r\n            \r\n            {\r\n              \"name\": \"Alice Wonder\",\r\n              \"email\": \"alice@example.com\",\r\n              \"age\": 28\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (201 Created):</strong>\r\n\r\n```\r\n            HTTP/1.1 201 Created\r\n            Location: http://localhost:5137/Users/4\r\n            \r\n            {\r\n              \"id\": 4,\r\n              \"name\": \"Alice Wonder\",\r\n              \"email\": \"alice@example.com\",\r\n              \"age\": 28,\r\n              \"createdAt\": \"2025-11-07T12:00:00Z\",\r\n              \"updatedAt\": null\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (400 Bad Request - Duplicate Email):</strong>\r\n\r\n```\r\n            {\r\n              \"message\": \"A user with this email already exists\"\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (400 Bad Request - Validation Error):</strong>\r\n\r\n```\r\n            {\r\n              \"type\": \"https://tools.ietf.org/html/rfc7231#section-6.5.1\",\r\n              \"title\": \"One or more validation errors occurred.\",\r\n              \"status\": 400,\r\n              \"errors\": {\r\n                \"Name\": [\"Name is required\"],\r\n                \"Age\": [\"Age must be between 1 and 150\"]\r\n              }\r\n            }\r\n            ```",
        "requestBody": {
          "description": "The user data to create. ID, CreatedAt, and UpdatedAt are auto-generated.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data or duplicate email",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieves a specific user by their unique identifier.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> GET\r\n\n\r\n  <strong>Route:</strong> /Users/{id}\r\n\n\r\n  <strong>Route Parameter:</strong> id (integer) - User's unique identifier\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            GET http://localhost:5137/Users/1\r\n            Accept: application/json\r\n            ```\r\n\n\r\n  <strong>Example Response (200 OK):</strong>\r\n\r\n```\r\n            {\r\n              \"id\": 1,\r\n              \"name\": \"John Doe\",\r\n              \"email\": \"john@example.com\",\r\n              \"age\": 30,\r\n              \"createdAt\": \"2025-11-07T10:46:00Z\",\r\n              \"updatedAt\": null\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (404 Not Found):</strong>\r\n\r\n```\r\n            {\r\n              \"message\": \"User with ID 999 not found\"\r\n            }\r\n            ```",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique ID of the user to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully returns the user",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User with the specified ID was not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Updates an existing user with new data.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> PUT\r\n\n\r\n  <strong>Route:</strong> /Users/{id}\r\n\n\r\n  <strong>Content-Type:</strong> application/json\r\n\n\r\n  <strong>Important:</strong> The ID in the URL must match the ID in the request body.\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            PUT http://localhost:5137/Users/1\r\n            Content-Type: application/json\r\n            \r\n            {\r\n              \"id\": 1,\r\n              \"name\": \"John Updated\",\r\n              \"email\": \"john.updated@example.com\",\r\n              \"age\": 31\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (204 No Content):</strong>\r\n\r\n```\r\n            HTTP/1.1 204 No Content\r\n            (No response body)\r\n            ```\r\n\n\r\n  <strong>Example Response (400 Bad Request - ID Mismatch):</strong>\r\n\r\n```\r\n            {\r\n              \"message\": \"ID mismatch between URL and request body\"\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (404 Not Found):</strong>\r\n\r\n```\r\n            {\r\n              \"message\": \"User with ID 1 not found\"\r\n            }\r\n            ```",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The updated user data. ID must match the URL parameter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User updated successfully (no response body)"
          },
          "400": {
            "description": "Invalid data or ID mismatch",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a user from the system.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> DELETE\r\n\n\r\n  <strong>Route:</strong> /Users/{id}\r\n\n\r\n  <strong>Warning:</strong> This operation is permanent and cannot be undone.\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            DELETE http://localhost:5137/Users/3\r\n            ```\r\n\n\r\n  <strong>Example Response (204 No Content):</strong>\r\n\r\n```\r\n            HTTP/1.1 204 No Content\r\n            (No response body)\r\n            ```\r\n\n\r\n  <strong>Example Response (404 Not Found):</strong>\r\n\r\n```\r\n            {\r\n              \"message\": \"User with ID 999 not found\"\r\n            }\r\n            ```",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the user to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully (no response body)"
          },
          "404": {
            "description": "User not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Users/average-age": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Calculates and returns the average age of all users in the system.",
        "description": "\n\r\n  <strong>HTTP Method:</strong> GET\r\n\n\r\n  <strong>Route:</strong> /Users/average-age\r\n\n\r\n  <strong>Use Case:</strong> Business analytics, dashboard metrics, reporting\r\n\n\r\n  <strong>How It Works:</strong>\r\n\r\n<list type=\"number\">\r\n  <item>Service layer retrieves all users from repository</item>\r\n  <item>Calculates average age using LINQ's Average() method</item>\r\n  <item>Returns 0 if no users exist (to avoid division by zero)</item>\r\n  <item>Rounds result to 2 decimal places for readability</item>\r\n</list>\r\n\n\r\n  <strong>Example Request:</strong>\r\n\r\n```\r\n            GET http://localhost:5137/Users/average-age\r\n            Accept: application/json\r\n            ```\r\n\n\r\n  <strong>Example Response (200 OK - With Users):</strong>\r\n\r\n```\r\n            {\r\n              \"averageAge\": 30.5,\r\n              \"totalUsers\": 4,\r\n              \"message\": \"Average calculated from 4 users\"\r\n            }\r\n            ```\r\n\n\r\n  <strong>Example Response (200 OK - No Users):</strong>\r\n\r\n```\r\n            {\r\n              \"averageAge\": 0,\r\n              \"totalUsers\": 0,\r\n              \"message\": \"No users in the system\"\r\n            }\r\n            ```\r\n\n\r\n  <strong>Calculation Example:</strong>\r\n\r\n\nIf users have ages: 30, 25, 35, 40\r\n\nAverage = (30 + 25 + 35 + 40) / 4 = 32.5",
        "responses": {
          "200": {
            "description": "Successfully calculated average age"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Gets a 5-day weather forecast.",
        "responses": {
          "200": {
            "description": "Returns the weather forecast data",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Health": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "User": {
        "required": [
          "age",
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the user.",
            "format": "int32",
            "example": 1
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string",
            "description": "Full name of the user.",
            "example": "John Doe"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "Email address of the user.",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "age": {
            "maximum": 150,
            "minimum": 1,
            "type": "integer",
            "description": "Age of the user in years.",
            "format": "int32",
            "example": 25
          },
          "createdAt": {
            "type": "string",
            "description": "UTC timestamp when the user was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "UTC timestamp when the user was last updated.",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents a user in the system.\r\nContains user identification, contact information, and metadata."
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "temperatureF": {
            "type": "integer",
            "description": "Temperature in Fahrenheit (computed).",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Represents a weather forecast for a single day."
      }
    }
  }
}